linked and got functionality working
