+++ public/assets/get-help.html
@@ -8,7 +8,7 @@
       rel="stylesheet"
       href="https://cdn.jsdelivr.net/npm/bulma@0.9.4/css/bulma.min.css"
     />
-    <link rel="stylesheet" href="/public/css/style.css" />
+    <link rel="stylesheet" href="../css/style.css" />
     <title>alluda!WelcomeHome</title>
   </head>
   <body class="hero background is-cover is-info" id="body">
