+++ public/index.html
@@ -1,46 +1,46 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang='en-US'>
   <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta charset='UTF-8' />
+    <meta http-equiv='X-UA-Compatible' content='IE=edge'/>
+    <meta name='viewport" content="width=device-width, initial-scale=1.0' />
     <link
-      rel="stylesheet"
-      href="https://cdn.jsdelivr.net/npm/bulma@0.9.4/css/bulma.min.css"
+      rel='stylesheet'
+      href='https://cdn.jsdelivr.net/npm/bulma@0.9.4/css/bulma.min.css'
     />
-    <link rel="stylesheet" href="/public/css/style.css" />
+    <link rel='stylesheet' href='../public/css/style.css' />
     <title>alluda!WelcomeHome</title>
   </head>
-  <body class="hero background is-cover" id="body">
+  <body class='hero background is-cover" id='body'>
     <header
-      class="background is-info hero-head ml-6 mr-2 title is-size-12 has-text-blue is-bold mt-6 is-info"
+      class='background is-info hero-head ml-6 mr-2 title is-size-12 has-text-blue is-bold mt-6 is-info'
     >
       alluda!
-      <nav class="rows">
+      <nav class='rows'>
         <!-- links go here -->
         <a
-          class="button is-focused is-normal is-responsive mb-2 ml-2 mt-2"
-          href="/get-help"
+          class='button is-focused is-normal is-responsive mb-2 ml-2 mt-2'
+          href='/get-help'
         >
           Certified Providers
         </a>
         <a
-          class="button is-focused is-normal is-responsive mb-2 ml-2 mt-2"
-          href="/media"
+          class='button is-focused is-normal is-responsive mb-2 ml-2 mt-2'
+          href='/media'
         >
           Media Resources
         </a>
         <a
-          class="button is-focused is-normal is-responsive mb-2 ml-2 mt-2"
-          href="/notes"
-          id="getNotes"
+          class='button is-focused is-normal is-responsive mb-2 ml-2 mt-2'
+          href='/notes'
+          id='getNotes'
         >
           Journal Log
         </a>
         <a
-          class="button is-focused is-normal is-responsive mb-2 ml-2 mt-2"
-          href="/about-us"
-          id="getAboutUs"
+          class='button is-focused is-normal is-responsive mb-2 ml-2 mt-2'
+          href='/about-us'
+          id='getAboutUs'
         >
           About Us
         </a>
@@ -91,6 +91,6 @@
         </p>
       </div>
     </footer>
-    <script href='/server.js'></script>
+    <script href='../server.js'></script>
   </body>
 </html>
