+++ public/js/notes.js
