+++ public/notesAssets/notes.html
@@ -1,80 +1,79 @@
 <!DOCTYPE html>
-<html lang="en-US">
+<html lang='en-US'>
   <head>
-    <meta charset="UTF-8" />
+    <meta charset='UTF-8' />
     <meta name='viewport" content="width=device-width, initial-scale=1.0' />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+    <meta http-equiv='X-UA-Compatible' content='ie=edge' />
     <title>alluda! Journal Log</title>
     <link
-      rel="stylesheet"
-      href="https://cdn.jsdelivr.net/npm/bulma@0.9.4/css/bulma.min.css"
+      rel='stylesheet'
+      href='https://cdn.jsdelivr.net/npm/bulma@0.9.4/css/bulma.min.css'
     />
     <link
-      rel="stylesheet"
-      href="https://cdnjs.cloudflare.com/ajax/libs/bootswatch/4.1.3/flatly/bootstrap.min.css"
+      rel='stylesheet'
+      href='https://cdnjs.cloudflare.com/ajax/libs/bootswatch/4.1.3/flatly/bootstrap.min.css'
     />
     <link
-      rel="stylesheet"
-      href="https://use.fontawesome.com/releases/v5.3.1/css/all.css"
-      integrity="sha384-mzrmE5qonljUremFsqc01SB46JvROS7bZs3IO2EmfFsd15uHvIt+Y8vEf7N7fWAU"
-      crossorigin="anonymous"
+      rel='stylesheet'
+      href='https://use.fontawesome.com/releases/v5.3.1/css/all.css'
+      integrity='sha384-mzrmE5qonljUremFsqc01SB46JvROS7bZs3IO2EmfFsd15uHvIt+Y8vEf7N7fWAU'
+      crossorigin='anonymous'
     />
-
-    <link rel="stylesheet" href="/public/css/style.css" />
+    <link rel='stylesheet' href='../css/notes.css'>
   </head>
 
   <body>
-    <nav class="navbar navbar-dark bg-dark">
-      <a class="navbar-brand" href="/">alluda!</a>
+    <nav class='navbar navbar-dark bg-dark'>
+      <a class='navbar-brand' href='/'>alluda!</a>
       <a
-        class="button is-focused is-normal is-responsive mb-2 ml-2 mt-2"
-        href="/get-help"
+        class='button is-focused is-normal is-responsive mb-2 ml-2 mt-2'
+        href='/get-help'
       >
         Certified Providers
       </a>
       <a
-        class="button is-focused is-normal is-responsive mb-2 ml-2 mt-2"
-        href="/media"
+        class='button is-focused is-normal is-responsive mb-2 ml-2 mt-2'
+        href='/media'
       >
         Media Resources
       </a>
       <a
-        class="button is-focused is-normal is-responsive mb-2 ml-2 mt-2"
-        href="/notes"
-        id="getNotes"
+        class='button is-focused is-normal is-responsive mb-2 ml-2 mt-2'
+        href='/notes'
+        id='getNotes'
       >
         Journal Log
       </a>
       <a
-        class="button is-focused is-normal is-responsive mb-2 ml-2 mt-2"
-        href="/about-us"
-        id="getAboutUs"
+        class='button is-focused is-normal is-responsive mb-2 ml-2 mt-2'
+        href='/about-us'
+        id='getAboutUs'
       >
         About Us
       </a>
-      <div class="icons">
-        <i class="fas fa-save has-text-black save-note"></i>
-        <i class="fas fa-plus has-text-black new-note"></i>
+      <div class='icons'>
+        <i class='fas fa-save has-text-black save-note'></i>
+        <i class='fas fa-plus has-text-black new-note'></i>
       </div>
     </nav>
-    <div class="container-fluid">
-      <div class="row">
-        <div class="col-4 list-container">
-          <div class="card">
-            <ul class="list-group"></ul>
+    <div class='container-fluid'>
+      <div class='row'>
+        <div class='col-4 list-container'>
+          <div class='card'>
+            <ul class='list-group'></ul>
           </div>
         </div>
-        <div class="col-8">
+        <div class='col-8'>
           <input
-            class="note-title"
-            placeholder="Note Title"
-            maxlength="28"
-            type="text"
+            class='note-title'
+            placeholder='Note Title'
+            maxlength='28'
+            type='text'
           />
-          <textarea class="note-textarea" placeholder="Note Text"></textarea>
+          <textarea class='note-textarea' placeholder='Note Text'></textarea>
         </div>
       </div>
     </div>
-    <script src="./notes.js"></script>
+    <script src='../js/notes.js'></script>
   </body>
 </html>
