+++ routes/apiRoutes/notesRoutes.js
@@ -1,4 +1,4 @@
-const router = require('express').Router();
+const router = require("express").Router();
 const {
   filterByQuery,
   findById,
@@ -6,16 +6,15 @@ const {
   validateNote,
   deleteNote,
   getAllNotes,
-} = require('../../public/notesAssets/noteTaker.js');
+} = require("../../lib/noteTaker.js");
 
-
-router.get('/notes', (req, res) => {
+router.get("/notes", (req, res) => {
   let results = getAllNotes().notes;
   console.log(results);
   return res.json(results);
 });
 
-router.get('/notes/:id', (req, res) => {
+router.get("/notes/:id", (req, res) => {
   const result = findById(req.params.id, getAllNotes().notes);
   if (result) {
     res.json(result);
@@ -24,20 +23,20 @@ router.get('/notes/:id', (req, res) => {
   }
 });
 
-router.post('/notes', (req, res) => {
+router.post("/notes", (req, res) => {
   // set id based on what the next index of the array will be
   console.log(getAllNotes().notes);
   req.body.id = getAllNotes().notes.length;
 
   if (!validateNote(req.body)) {
-    res.status(400).send('The notes are not properly formatted.');
+    res.status(400).send("The notes are not properly formatted.");
   } else {
     const note = createNewNote(req.body, getAllNotes().notes);
     res.json(note);
   }
 });
 
-router.delete('/notes/:id', (req, res) => {
+router.delete("/notes/:id", (req, res) => {
   const result = deleteNote(req.params.id);
   if (result) {
     res.json(result);
